streamlit_app/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
opencc-python-reimplemented>=0.1.7


